move mesh.cpp/hpp back to serac_physics
