+++ CMakeLists.txt
@@ -195,7 +195,7 @@ configure_file(
 #------------------------------------------------------------------------------
 set(exported_targets
     serac_infrastructure
-    serac_mesh
+    serac_mesh_utils
     serac_numerics
     serac_physics)
