+++ src/serac/infrastructure/tests/error_handling.cpp
@@ -12,7 +12,7 @@
 #include "serac/infrastructure/cli.hpp"
 #include "serac/infrastructure/application_manager.hpp"
 #include "serac/mesh_utils/mesh_utils.hpp"
-#include "serac/physics/mesh/mesh.hpp"
+#include "serac/physics/mesh.hpp"
 #include "serac/physics/boundary_conditions/boundary_condition.hpp"
 #include "serac/numerics/equation_solver.hpp"
 #include "serac/serac_config.hpp"
