+++ src/serac/numerics/CMakeLists.txt
@@ -23,7 +23,7 @@ set(numerics_sources
     petsc_solvers.cpp
     )
 
-set(numerics_depends serac_infrastructure serac_functional serac_mesh)
+set(numerics_depends serac_infrastructure serac_functional)
 
 blt_add_library(
     NAME        serac_numerics
