+++ src/serac/numerics/tests/test_trust_region_solver.cpp
@@ -10,7 +10,7 @@
 
 #include "mfem.hpp"
 #include "serac/mesh_utils/mesh_utils.hpp"
-#include "serac/physics/mesh/mesh.hpp"
+#include "serac/physics/mesh.hpp"
 #include "serac/physics/state/state_manager.hpp"
 
 #include "serac/numerics/functional/functional.hpp"
