+++ src/serac/physics/CMakeLists.txt
@@ -8,10 +8,10 @@ add_subdirectory(materials)
 add_subdirectory(state)
 add_subdirectory(boundary_conditions)
 add_subdirectory(contact)
-add_subdirectory(mesh)
 
 set(physics_sources
     base_physics.cpp
+    mesh.cpp
     solid_mechanics.cpp
     heat_transfer_input.cpp
     solid_mechanics_input.cpp
@@ -20,6 +20,7 @@ set(physics_sources
 
 set(physics_headers
     base_physics.hpp
+    mesh.hpp
     common.hpp
     fit.hpp
     heat_transfer.hpp
@@ -43,7 +44,6 @@ set(physics_depends
     serac_contact
     serac_numerics
     serac_physics_materials
-    serac_mesh
     )
 
 blt_add_library(
