+++ src/serac/physics/functional_objective.hpp
@@ -13,7 +13,7 @@
 #pragma once
 
 #include "serac/physics/scalar_objective.hpp"
-#include "serac/physics/mesh/mesh.hpp"
+#include "serac/physics/mesh.hpp"
 #include "serac/numerics/functional/shape_aware_functional.hpp"
 #include "serac/physics/state/finite_element_state.hpp"
 #include "serac/physics/state/finite_element_dual.hpp"
