+++ src/serac/physics/state/tests/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 
-set(state_test_depends serac_state serac_physics_materials serac_functional serac_mesh_utils gtest)
+set(state_test_depends serac_state serac_physics_materials serac_functional gtest)
 
 set(state_test_sources
     test_finite_element_dual.cpp
