+++ src/serac/physics/tests/test_functional_residual.cpp
@@ -8,7 +8,7 @@
 #include "mfem.hpp"
 #include "serac/infrastructure/application_manager.hpp"
 #include "serac/mesh_utils/mesh_utils.hpp"
-#include "serac/physics/mesh/mesh.hpp"
+#include "serac/physics/mesh.hpp"
 #include "serac/physics/state/state_manager.hpp"
 
 #include "serac/physics/functional_residual.hpp"
