+++ src/serac/physics/tests/test_kinematic_objective.cpp
@@ -12,7 +12,7 @@
 #include "serac/physics/state/state_manager.hpp"
 #include "serac/mesh_utils/mesh_utils.hpp"
 #include "serac/physics/materials/solid_material.hpp"
-#include "serac/physics/mesh/mesh.hpp"
+#include "serac/physics/mesh.hpp"
 #include "serac/physics/common.hpp"
 #include "mfem.hpp"
 #include "serac/physics/tests/physics_test_utils.hpp"
